[#753] Move QuestionField* components to own file
